test: Split workflow-actions.spec.ts into focused test files
